qwait-arima-service/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
statsmodels==0.14.0
scipy==1.11.2
python-dateutil==2.8.2
opencv-python==4.8.0
opencv-contrib-python==4.8.0
Pillow==10.0.0
pytz==2023.3
scikit-learn==1.3.0